:truck: move config_from_module & config_from_static_file in environment
